fuse-adaptor: move fuse adaptor source code file to special folder
