+++ include/photon/fs/fuse_adaptor.h
@@ -1,1 +1,1 @@
-../../../fs/fuse_adaptor.h
+../../../fs/fuse_adaptor/fuse_adaptor.h
