--- include/photon/io/fuse-adaptor.h
@@ -1,1 +1,0 @@
-../../../fs/fuse_adaptor.h
